full finish on saved texts modal
