+++ frontend/src/html/popups.html
@@ -467,7 +467,7 @@
     <div class="button apply">ok</div>
   </div>
 </dialog>
-<dialog id="savedTextsPopupWrapper" class="modalWrapper hidden">
+<dialog id="savedTextsModal" class="modalWrapper hidden">
   <div class="modal">
     <div class="title">Saved texts</div>
     <div class="list"></div>
