+++ frontend/src/styles/popups.scss
@@ -189,7 +189,7 @@
   }
 }
 
-#savedTextsPopupWrapper {
+#savedTextsModal {
   .modal {
     max-width: 500px;
     .list {
